# Conflicts:
#	src/main/java/com/xsis/bootcamp/model/Status.java
#	src/main/java/com/xsis/bootcamp/model/common/CommonModel.java

+++ src/main/java/com/xsis/bootcamp/service/MroleService.java
@@ -6,7 +6,7 @@ import com.xsis.bootcamp.model.Mrole;
 
 public interface MroleService {
 
-	public Mrole get(Long idRole) throws Exception;
+	public Mrole get(Integer idRole) throws Exception;
 
 	public abstract void insert(Mrole mrole) throws Exception;
